refactor balance management
